[Fix/#101] clean file and folder
