+++ Onbom/Onbom/Model/Address.swift
@@ -1,0 +1,12 @@
+//
+//  Address.swift
+//  Onbom
+//
+//  Created by moon on 10/27/23.
+//
+
+
+struct Address {
+    var cityAddress: String = ""
+    var detailAddress: String = ""
+}
