--- Onbom/Onbom/View/ContentView.swift
@@ -1,26 +1,0 @@
-//
-//  ContentView.swift
-//  Onbom
-//
-//  Created by moon on 2023/09/22.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundColor(.accentColor)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
